Hv: Update interrupt/exception handle behaviors
In Fusa scenario, the interrupts shall be disabled in vmx root mode
except NMI and defensive programming shall be adopted. The unmask-able
NMI shall be inject to guest vm if happend. Besides, the exceptions
are also unexpected in vmx root mode.

The handle_nmi() is modified to inject NMI to guest vm and
dispatch_exception() is modified to call panic() for defensive
programming according to SAS design.

v1->v2:
Modify the behavior of handle_nmi() to avoid inappropriate invocation of
upper layer function in the lower layer.

v2->v3:
*Regarding the consistency of coding style, just take the invocation of
vcpu_queue_exception of virq.c as exception and roll back to the first version
to avoid excursive changes.
*Combine the deletion of the unused affix in this patch.

Signed-off-by: Junjun Shan <junjun.shan@intel.com>