hv: ptdev: restore the template for module vp-dm.ptirq
In commit 7e7aeb6 ("hv: ptirq: fix to meet the SAS requirement"), common/ptdev.c
was removed because vmsis are not associated with any physical irq. However the
template for module vp-dm.ptirq happened to be in that file. The file removal
caused the module to be undefined throughout the code, leading to some trouble
during module design generation.

This patch restores the template by putting it to arch/x86/guest/assign.c.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>