hwmgmt.page: refine some given parameters
Refine the given parameters to the following functions in order to
save the effort to explain the tricky logic for every function
definition and function invocation in MDS.
 - modify_or_del_pte
 - modify_or_del_pde
 - modify_or_del_pdpte
 - add_pte
 - add_pde
 - add_pdpte

Take the definition of 'modify_or_del_pde' as an example:
'modify_or_del_pte' is invoked inside to handle the PTEs in one
page table. So, the actual memory region to be handled in each
iteration is [vaddr, vaddr_end_each_iter).
vaddr_end_each_iter =
   (vaddr_next < vaddr_end) ? vaddr_next : vaddr_end;
If vaddr_end is larger than vaddr_next, the memory region specified by
[vaddr_next, vaddr_end) is handled in next iterations rather than the
current one.

Signed-off-by: Shiqing Gao <shiqing.gao@intel.com>