treewide: fix space issues and unnecessary symbols
Lists are expected to be aligned using spaces, and macro definitions shall
have the hash symbol ("#") as the first character of the line. This patch
adds or removes the white spaces according to the above rules.

Also, values to @asinteger commands do not need a closing brace. This patch
also remove this character if there is any.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>