cpu: add comments to the symbols defined in assembly or link script
The symbols defined in assembly or the link script do not have their
definition within the scope of files Doxygen can analyze. This patch
describes their intention and usage at their declarations instead so
that they can be properly explained in the module design.

The descriptions mostly come from SAS which is also referred when more
info is available there.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>