treewide: explain undescribed magic numbers
According to rule C-EP-22 in the coding guideline, magic numbers shall be
used either by wrapping up as macros or documenting their meaning when they
are used. This patch adds explanations to the magic numbers that are not
properly described yet.

These are identified during the checks against rule 2.1 in code review.

Tracked-On: ACCB-134
Signed-off-by: Junjie Mao <junjie.mao@intel.com>