pci: split initializers to different lines
According to rule C-CS-02 in the coding guideline, each line shall have at
most one statement. While initializers are not strictly considered as
statements, it is a good practice to have only one initializer per source
line for readability. This patch fixes a one-liner initialization that
violates this rule.

This issue identified during the checks against rule 2.3 in code review.

Tracked-On: ACCB-134
Signed-off-by: Junjie Mao <junjie.mao@intel.com>