HV: Fix PCI bugs
Tracked on CCB 133.
ACRN-6750
Add logic to limit the behavior of the PCI MSI configuration, while the
interrupt vector is not in the range [0x10, 0xfe], the PCI MSI interrupt
is disabled.

ACRN-6472
Set the initial value of the field of the hostbridge PCI configuration
interrupt line to FFH.

ACRN-6495
Set the initial value of the field of the PCI configuration interrupt
pin of any PCI assigned PCIe device to 0H.

ACRN-6574:
Set the initial value of the field of the PCI configuration guest
message address register of any PCI assigned PCIe device to FEE00000H.

ACRN-6575:
Set the initial value of the bit[15:14], bit[10:8] of the message data
register of an assigned PCIe device to 0.

ACRN-6576 & ACRN-6578 & ACRN-6579 & ACRN-6278 & ACRN-6581:
Set read-only bit to mask the value that will be written to PCI MSI
configuration space.

Signed-off-by: huihuang.shi <huihuang.shi@intel.com>