vmx_io: always overwrite guest CR2 with 0 upon invalid accesses
As has been decided by the CCB, it is of no sense to set guest CR2 to the
faulting guest physical address when a vCPU attempts to access one that
maps to none.

This patch overwrites the guest CR2 with 0 upon such invalid accesses, as
is recorded as item 79 of ACCB-134.

Tracked-On: ACCB-134
Signed-off-by: Junjie Mao <junjie.mao@intel.com>