pagetable: fix a potential dead loop
There is a dead loop in mmu_modify_or_del() which can be triggered by a
vCPU attempting to fetch instructions from a guest physical address that
maps to none. This patch fixes the dead loop and adds thorough checks on
the faulting address upon EPT violation to make sure that we only grant
execution permission to addresses that map to memory.

This is recorded as item 81 in ACCB-134.

Tracked-On: ACCB-134
Signed-off-by: Junjie Mao <junjie.mao@intel.com>