pagetable: update the policy of mapping unaligned regions
The policy of mapping unaligned regions is not consistent in its current
shape: the base address of a region is rounded up (i.e. restrictive), while
the end is rounded up as well (i.e. permissive).

This patch unifies the policy to be permissive, i.e. the base address is
rounded down while the the end is rounded up.

This is tracked as item 77 of ACCB-134.

Tracked-On: ACCB-134
Signed-off-by: Junjie Mao <junjie.mao@intel.com>