vmx_io: apply defensive actions on range check errors
This patch adds the general handling of range check errors on VM exit
qualifications, i.e. call panic() when it is triggered by the safety VM or
shutdown the VM otherwise. The size of I/O instructions is now checked to
make sure it is valid.

This is recorded as item 65 in ACCB-134.

Tracked-On: ACCB-134
Signed-off-by: Junjie Mao <junjie.mao@intel.com>